Support multiple split accounts per transaction (#51)
Do not go through the postings on split accounts per posting of another
tax slip relevant account.
This way we previously needed to remove the transaction once processed.
This was to avoid that when a transaction has multiple accounts and a
split account, it would be counted multiple times.

By going through the split accounts instead, we do not count it multiple
times anymore, so we no longer need to remove it and can handle multiple
split accounts per transaction.